activities-backend/requirements.txt
Django
djangorestframework
django-allauth
django-extensions
django-environ
psycopg
pillow
qrcode
django-cors-headers
rdflib
